Move clang matcher in rule_src_parser
